Refactor: implemented CollectorMode for insight collector and made it behave according to the mode
