+++ pkg/app/ops/insightcollector/applications.go
@@ -1,0 +1,33 @@
+// Copyright 2021 The PipeCD Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package insightcollector
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/pipe-cd/pipe/pkg/model"
+)
+
+// ProcessApplications collects data from applicationstore
+func ProcessApplications(_ context.Context) error {
+	return fmt.Errorf("not implemented yet")
+}
+
+// collectApplicationCount collects application count data.
+func (i *InsightCollector) collectApplicationCount(_ context.Context, _ []*model.Application, _ time.Time) error {
+	return fmt.Errorf("not implemented yet")
+}
