+++ pkg/app/ops/insightcollector/deployments_test.go
